+++ ci/licenses_golden/licenses_flutter
@@ -2412,9 +2412,6 @@ FILE: ../../../flutter/shell/platform/windows/keyboard_utils_unittests.cc
 FILE: ../../../flutter/shell/platform/windows/platform_handler.cc
 FILE: ../../../flutter/shell/platform/windows/platform_handler.h
 FILE: ../../../flutter/shell/platform/windows/platform_handler_unittests.cc
-FILE: ../../../flutter/shell/platform/windows/platform_handler_win32.cc
-FILE: ../../../flutter/shell/platform/windows/platform_handler_win32.h
-FILE: ../../../flutter/shell/platform/windows/platform_handler_win32_unittests.cc
 FILE: ../../../flutter/shell/platform/windows/public/flutter_windows.h
 FILE: ../../../flutter/shell/platform/windows/sequential_id_generator.cc
 FILE: ../../../flutter/shell/platform/windows/sequential_id_generator.h
