+++ shell/platform/windows/BUILD.gn
@@ -82,8 +82,6 @@ source_set("flutter_windows_source") {
     "keyboard_utils.h",
     "platform_handler.cc",
     "platform_handler.h",
-    "platform_handler_win32.cc",
-    "platform_handler_win32.h",
     "sequential_id_generator.cc",
     "sequential_id_generator.h",
     "settings_plugin.cc",
@@ -184,7 +182,6 @@ executable("flutter_windows_unittests") {
     "keyboard_unittests.cc",
     "keyboard_utils_unittests.cc",
     "platform_handler_unittests.cc",
-    "platform_handler_win32_unittests.cc",
     "sequential_id_generator_unittests.cc",
     "settings_plugin_unittests.cc",
     "system_utils_unittests.cc",
